pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = ["fastapi (>=0.115.7,<0.116.0)", "langchain-ollama (>=0.2.2,<0.3.0)", "langchain (>=0.3.15,<0.4.0)", "langchain-openai (>=0.3.2,<0.4.0)", "tortoise-orm (>=0.24.0,<0.25.0)", "psycopg2-binary (>=2.9.10,<3.0.0)", "python-multipart (>=0.0.20,<0.0.21)", "jinja2 (>=3.1.5,<4.0.0)", "pydantic-settings (>=2.7.1,<3.0.0)", "pyjwt (>=2.10.1,<3.0.0)", "passlib (>=1.7.4,<2.0.0)", "asyncpg (>=0.30.0,<0.31.0)", "uvicorn[standard] (>=0.34.0,<0.35.0)", "aerich (>=0.8.1,<0.9.0)"]

[[project.authors]]
name = "shkrobik2017"
email = "shkrobik2017@gmail.com"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
